♻️ refactor: 현재 사용되지 않는 lib/github의 utils에서 데이터 페칭 관리하기
