+++ src/app/posting/[slug]/page.tsx
@@ -4,61 +4,13 @@ import remarkGfm from 'remark-gfm';
 import rehypeSanitize from 'rehype-sanitize';
 import { MdxStyle } from '../components/MdxStyle';
 import { notFound } from 'next/navigation';
+import { getMarkdownContent } from '../../../lib/github';
 
 interface Params {
 	slug: string;
 }
 interface Props {
 	params: Promise<Params>;
-}
-
-/**
- * Github 환경 변수 체크
- */
-function requireGitEnv() {
-	const GitEnv = {
-		GITHUB_OWNER: process.env.GITHUB_OWNER,
-		GITHUB_REPO: process.env.GITHUB_REPO,
-		GITHUB_TOKEN: process.env.GITHUB_TOKEN,
-	};
-	const missing = Object.entries(GitEnv).filter(([, v]) => !v?.trim());
-	if (missing.length) {
-		throw new Error(
-			`환경 변수 설정 오류: ${missing.join(', ')} 누락되었습니다.`,
-		);
-	}
-	const {
-		GITHUB_OWNER: owner,
-		GITHUB_REPO: repo,
-		GITHUB_TOKEN: token,
-	} = GitEnv as {
-		GITHUB_OWNER: string;
-		GITHUB_REPO: string;
-		GITHUB_TOKEN: string;
-	};
-
-	return { owner, repo, token };
-}
-
-/**
- * Github API가 반환하는 Base64 형태의 파일을 utf-8로 반환
- */
-async function getMarkdownContent(slug: string) {
-	const { owner, repo, token } = requireGitEnv();
-
-	const repoUrl = `https://api.github.com/repos/${owner}/${repo}/contents/til/${slug}.md`;
-	const res = await fetch(repoUrl, {
-		headers: { Authorization: `token ${token}` },
-		cache: 'no-cache',
-	});
-	if (res.status === 404) return null;
-	if (!res.ok) {
-		throw new Error(`markdown fetching 오류: ${res.status} ${res.statusText}`);
-	}
-
-	const { content } = await res.json();
-	const markdown = Buffer.from(content, 'base64').toString('utf-8');
-	return markdown;
 }
 
 /**
